+++ example/lib/ui/app.dart
@@ -13,25 +13,24 @@
 // limitations under the License.
 
 import 'package:flutter/material.dart';
-import 'package:push_demo/ui/message_screen.dart';
+import 'package:push_demo/notification/messaging_service.dart';
+import 'package:push_demo/ui/main_screen.dart';
 import 'package:push_notification/push_notification.dart';
 
 class MyApp extends StatelessWidget {
   final PushHandler _pushHandler;
+  final MessagingService _messagingService;
 
-  const MyApp(this._pushHandler, {super.key});
+  const MyApp(this._pushHandler, this._messagingService, {super.key});
 
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      navigatorObservers: [
-        PushObserver(),
-      ],
+      navigatorObservers: [PushObserver()],
       title: 'Push demo',
-      theme: ThemeData(
-        primarySwatch: Colors.indigo,
-      ),
-      home: MessageScreen(_pushHandler),
+      theme: ThemeData(primarySwatch: Colors.indigo),
+      home: MainScreen(
+          pushHandler: _pushHandler, messagingService: _messagingService),
     );
   }
 }